+++ test/gpui/schema_test.exs
@@ -6,6 +6,9 @@ defmodule GPUI.SchemaTest do
              :div,
              :button,
              :ui_button,
+             :ui_progress,
+             :ui_file_picker,
+             :ui_copy_button,
              :ui_popover,
              :ui_popover_trigger,
              :ui_popover_content,
@@ -43,6 +46,9 @@ defmodule GPUI.SchemaTest do
     assert :"phx-search" in GPUI.Schema.events()
     assert :"phx-release" in GPUI.Schema.events()
     assert :"phx-select" in GPUI.Schema.events()
+    assert :ui_progress in GPUI.Schema.identified_tags()
+    assert :ui_file_picker in GPUI.Schema.identified_tags()
+    assert :ui_copy_button in GPUI.Schema.identified_tags()
     assert :ui_select in GPUI.Schema.identified_tags()
     assert :ui_popover in GPUI.Schema.identified_tags()
     assert :ui_tooltip in GPUI.Schema.identified_tags()